Update README.md
Hi Tim, just a few minor changes in your instructions.  The biggest one is making sure the .sh file is executable.

I also made some changes to the data in the bucket to make it work with my demo script for the workshop.  I'll get those to you in a few days.

--Oleg